use this commit to revert to non-many to many

+++ application.py
@@ -1,9 +1,23 @@
-from flask import Flask, render_template, request, redirect, url_for, jsonify
+from flask import Flask, render_template, request, redirect, url_for, jsonify, flash
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item
+from database_setup import Base, Category, Item, User
+from flask import session as login_session
+import random, string
 
+
+# IMPORTS FOR THIS STEP
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
 app = Flask(__name__)
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Item Catalog App"
 
 engine = create_engine('sqlite:///itemcatalog.db')
 Base.metadata.bind = engine
@@ -19,19 +33,160 @@ def categoryItemsJSON(category_id):
         category_id=category_id).all()
     return jsonify(Items=[i.serialize for i in items])
 
+@app.route('/items/<int:item_id>/JSON')
+def singleItemJSON(item_id):
+    item = session.query(Item).filter_by(
+        id=item_id).one()
+    return jsonify(Item=item.serialize)
 
-# # ADD JSON ENDPOINT HERE
-# @app.route('/restaurants/<int:restaurant_id>/menu/JSON')
-# def menuItemJSON(restaurant_id, menu_id):
-#     menuItem = session.query(MenuItem).filter_by(id=menu_id).one()
-#     return jsonify(MenuItem=menuItem.serialize)
+
+# ADD JSON ENDPOINT HERE
+@app.route('/catalog/JSON')
+def allCatalogItems():
+    allItems = session.query(Item).all()
+    return jsonify(Items=[i.serialize for i in allItems])
+
+#LOGIN STUFF
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
 
 
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = '  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"' 
+    output+= 'integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">'
+    output += '<link rel=stylesheet type=text/css href={{ url_for("static", filename="styles.css") }}">'
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 100px; height: 100px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("You are now logged in as %s" % login_session['username'])
+    return output
+
+
+@app.route('/logout')
+def gdisconnect():
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: '
+    print login_session['username']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        flash("You are now logged out")
+        return redirect('/')
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+#APP ROUTING STUFF
 @app.route('/')
 def displayCategories():
     categories = session.query(Category).all()
     return render_template(
         'catalog.html', categories = categories)
+
+@app.route('/categories/new')
+def newCategory():
+    items = session.query(Item).all()
+    user_id = login_session['user_id']
+    return render_template('newcategory.html', items = items, user_id=user_id)
 
 @app.route('/categories/<int:category_id>/items')
 def displayItemsInCategory(category_id):
@@ -49,11 +204,12 @@ def displayIndividualItem(item_id):
 
 @app.route('/items/new', methods=['GET', 'POST'])
 def newItem():
-
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         category_id=request.form['category_id']
         newItem = Item(name=request.form['name'], description=request.form[
-                           'description'], category_id=request.form['category_id'])
+                           'description'], category_id=request.form['category_id'], user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         return redirect(url_for('displayItemsInCategory', category_id=category_id))
@@ -65,7 +221,12 @@ def newItem():
 @app.route('/items/<int:item_id>/edit',
            methods=['GET', 'POST'])
 def editItem(item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     editedItem = session.query(Item).filter_by(id=item_id).one()
+    if editedItem.user_id != login_session['user_id']:
+        flash("You cannot edit this item because you did not create it.")
+        return redirect(url_for('displayItemsInCategory', category_id=editedItem.category_id))
     categories = session.query(Category).all()
     if request.method == 'POST':
         if request.form['name']:
@@ -86,7 +247,12 @@ def editItem(item_id):
 @app.route('/items/<int:item_id>/delete',
            methods=['GET', 'POST'])
 def deleteItem(item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     itemToDelete = session.query(Item).filter_by(id=item_id).one()
+    if itemToDelete.user_id != login_session['user_id']:
+        flash("You cannot delete this item because you did not create it.")
+        return redirect(url_for('displayItemsInCategory', category_id=itemToDelete.category_id))
     if request.method == 'POST':
         category_id = itemToDelete.category_id
         session.delete(itemToDelete)
@@ -96,6 +262,30 @@ def deleteItem(item_id):
         return render_template('delete.html', itemToDelete=itemToDelete)
 
 
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
 if __name__ == '__main__':
+    app.secret_key = 'v0RNg7e4ft3VUR61D2MebuIq'
     app.debug = True
+    logged_in_user = None
     app.run(host='0.0.0.0', port=5000)
